o/snapstate: change tests that omitted channel in actions when refreshing without providing any revision options
Previously, when refreshing snaps without a channel specifically set by
the caller, we didn't set the channel in the snap actions. The code is
simpler if we just always send the channel in the action, and default to
using the channel that is already set. We might not want this change, as
I am not sure if it has any unwanted side effects.