o/snapstate: allow link-component tasks to be added to pre or post-reboot chains
This fixes a bug that resulted in tasks existing between two edges that
were intended to be adjacent, under some rare conditions.

In the original implementation, link-component tasks were being inserted
between two edges, the MaybeRebootEdge and the MaybeRebootWaitEdge.
These edges normally point to link-snap and auto-connect. However, these
edges are advanced to prepare-kernel-modules-components and
discard-old-kernel-snap-setup when kernel module components are involved
in the refresh. Thus, this issue is normally covered up because kernel
module components are currently the only usage of components by
essential snaps.

To summarize, the placement of MaybeRebootWaitEdge is wrong when
refreshing an essential snap that uses components that are not kernel
module components. Under these conditions in the current implementation,
MaybeRebootWaitEdge points to auto-connect. It actually should point to
the first link-component task, if there are any.

This issue is further complicated by the handling of "finish-restart" on
a task. Currently, either auto-connect or discard-old-kernel-snap-setup
have "finish-restart" set to true on them. This commit does not change
that, despite moving the MaybeRebootWaitEdge to link-component, under
some conditions.